+++ test/ParsecSpec.hs
@@ -13,110 +13,100 @@ spec = parallel $ do
     parseArgs "" `shouldBe` emptyArgTree
     parseArgs "bad-cmd" `shouldBe` emptyArgTree
 
-  describe "create expr" $ do
+  describe "add expr" $ do
     it "invalid args" $ do
-      parseArgs "create" `shouldBe` emptyArgTree
-      parseArgs "create --json" `shouldBe` emptyArgTree
+      parseArgs "add" `shouldBe` emptyArgTree
+      parseArgs "add --json" `shouldBe` emptyArgTree
 
     it "with desc" $ do
-      parseArgs "create desc1 desc2    desc3" `shouldBe` emptyArgTree
-        { _cmd  = "create"
-        , _desc = "desc1 desc2 desc3"
-        }
+      parseArgs "add desc1 desc2    desc3"
+        `shouldBe` emptyArgTree { _cmd = "add", _desc = "desc1 desc2 desc3" }
       parseArgs "add + :desc [desc desc{}" `shouldBe` emptyArgTree
-        { _cmd  = "create"
+        { _cmd  = "add"
         , _desc = "+ :desc [desc desc{}"
         }
 
     it "with tags" $ do
-      parseArgs "create +tag1 +tag2"
-        `shouldBe` emptyArgTree { _cmd = "create", _tags = ["tag1", "tag2"] }
+      parseArgs "add +tag1 +tag2"
+        `shouldBe` emptyArgTree { _cmd = "add", _tags = ["tag1", "tag2"] }
       parseArgs "add +tag1   -tag1  +tag2" `shouldBe` emptyArgTree
-        { _cmd  = "create"
+        { _cmd  = "add"
         , _desc = "-tag1"
         , _tags = ["tag1", "tag2"]
         }
 
     it "with due" $ do
-      parseArgs "create  :10:20   " `shouldBe` emptyArgTree
-        { _cmd = "create"
+      parseArgs "add  :10:20   " `shouldBe` emptyArgTree
+        { _cmd = "add"
         , _due = Just $ ArgDate 10 0 0 20 0
         }
       parseArgs "add    ::20" `shouldBe` emptyArgTree
-        { _cmd = "create"
+        { _cmd = "add"
         , _due = Just $ ArgDate 0 0 0 20 0
         }
 
     it "with different order" $ do
-      let expectedTree = emptyArgTree { _cmd  = "create"
+      let expectedTree = emptyArgTree { _cmd  = "add"
                                       , _desc = "desc"
                                       , _tags = ["tag"]
                                       , _due  = Just $ ArgDate 0 0 0 20 0
                                       , _opts = ArgOpts { _json = True }
                                       }
-      parseArgs "create   desc +tag ::20 --json" `shouldBe` expectedTree
-      parseArgs "create +tag   desc --json ::20" `shouldBe` expectedTree
-      parseArgs "add ::20 --json   desc +tag" `shouldBe` expectedTree
+      parseArgs "a   desc +tag ::20 --json" `shouldBe` expectedTree
+      parseArgs "a +tag   desc --json ::20" `shouldBe` expectedTree
+      parseArgs "a ::20 --json   desc +tag" `shouldBe` expectedTree
 
-  describe "update expr" $ do
+  describe "edit expr" $ do
     it "invalid args" $ do
-      parseArgs "update" `shouldBe` emptyArgTree
-      parseArgs "update bad-id" `shouldBe` emptyArgTree
-      parseArgs "update 1" `shouldBe` emptyArgTree
+      parseArgs "edit" `shouldBe` emptyArgTree
+      parseArgs "edit bad-id" `shouldBe` emptyArgTree
+      parseArgs "e 1" `shouldBe` emptyArgTree
 
     it "with desc" $ do
-      parseArgs "update 1 desc" `shouldBe` emptyArgTree { _cmd  = "update"
-                                                        , _ids  = [1]
-                                                        , _desc = "desc"
-                                                        }
+      parseArgs "edit 1 desc"
+        `shouldBe` emptyArgTree { _cmd = "edit", _ids = [1], _desc = "desc" }
 
     it "with tags" $ do
-      parseArgs "update  1   +tag" `shouldBe` emptyArgTree { _cmd  = "update"
-                                                           , _ids  = [1]
-                                                           , _tags = ["tag"]
-                                                           }
+      parseArgs "edit  1   +tag"
+        `shouldBe` emptyArgTree { _cmd = "edit", _ids = [1], _tags = ["tag"] }
       parseArgs "edit 2   +tag -tag   +tag2" `shouldBe` emptyArgTree
-        { _cmd  = "update"
+        { _cmd  = "edit"
         , _ids  = [2]
         , _tags = ["tag2"]
         }
 
     it "with due" $ do
       parseArgs "edit   2 :10:20 " `shouldBe` emptyArgTree
-        { _cmd = "update"
+        { _cmd = "edit"
         , _ids = [2]
         , _due = Just $ ArgDate 10 0 0 20 0
         }
 
     it "with opt" $ do
-      parseArgs "update 1 desc +tag --json" `shouldBe` emptyArgTree
-        { _cmd  = "update"
+      parseArgs "e 1 desc +tag --json" `shouldBe` emptyArgTree
+        { _cmd  = "edit"
         , _desc = "desc"
         , _tags = ["tag"]
         , _ids  = [1]
         , _opts = ArgOpts True
         }
 
-  describe "replace expr" $ do
+  describe "set expr" $ do
     it "invalid args" $ do
-      parseArgs "replace" `shouldBe` emptyArgTree
-      parseArgs "replace bad-id" `shouldBe` emptyArgTree
-      parseArgs "replace 1" `shouldBe` emptyArgTree
-      parseArgs "replace 1 --json" `shouldBe` emptyArgTree
+      parseArgs "set" `shouldBe` emptyArgTree
+      parseArgs "set bad-id" `shouldBe` emptyArgTree
+      parseArgs "s 1" `shouldBe` emptyArgTree
+      parseArgs "s 1 --json" `shouldBe` emptyArgTree
 
     it "with desc" $ do
-      parseArgs "replace 1 desc" `shouldBe` emptyArgTree { _cmd  = "replace"
-                                                         , _ids  = [1]
-                                                         , _desc = "desc"
-                                                         }
+      parseArgs "set 1 desc"
+        `shouldBe` emptyArgTree { _cmd = "set", _ids = [1], _desc = "desc" }
 
     it "with tags" $ do
-      parseArgs "replace  1   +tag" `shouldBe` emptyArgTree { _cmd  = "replace"
-                                                            , _ids  = [1]
-                                                            , _tags = ["tag"]
-                                                            }
-      parseArgs "set 2   +tag -tag   +tag2" `shouldBe` emptyArgTree
-        { _cmd  = "replace"
+      parseArgs "set  1   +tag"
+        `shouldBe` emptyArgTree { _cmd = "set", _ids = [1], _tags = ["tag"] }
+      parseArgs "s 2   +tag -tag   +tag2" `shouldBe` emptyArgTree
+        { _cmd  = "set"
         , _ids  = [2]
         , _desc = "-tag"
         , _tags = ["tag", "tag2"]
@@ -124,14 +114,14 @@ spec = parallel $ do
 
     it "with due" $ do
       parseArgs "set   2 :10:20 " `shouldBe` emptyArgTree
-        { _cmd = "replace"
+        { _cmd = "set"
         , _ids = [2]
         , _due = Just $ ArgDate 10 0 0 20 0
         }
 
     it "with opt" $ do
-      parseArgs "replace 1 desc +tag --json" `shouldBe` emptyArgTree
-        { _cmd  = "replace"
+      parseArgs "set 1 desc +tag --json" `shouldBe` emptyArgTree
+        { _cmd  = "set"
         , _desc = "desc"
         , _tags = ["tag"]
         , _ids  = [1]
@@ -147,10 +137,10 @@ spec = parallel $ do
       parseArgs "start 1" `shouldBe` emptyArgTree { _cmd = "start", _ids = [1] }
       parseArgs "start 1 2"
         `shouldBe` emptyArgTree { _cmd = "start", _ids = [1, 2] }
-      parseArgs "start 1 --json" `shouldBe` emptyArgTree { _cmd  = "start"
-                                                         , _ids  = [1]
-                                                         , _opts = ArgOpts True
-                                                         }
+      parseArgs "+ 1 --json" `shouldBe` emptyArgTree { _cmd  = "start"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "stop expr" $ do
     it "invalid args" $ do
@@ -159,10 +149,10 @@ spec = parallel $ do
 
     it "with id" $ do
       parseArgs "stop 1" `shouldBe` emptyArgTree { _cmd = "stop", _ids = [1] }
-      parseArgs "stop 1 --json" `shouldBe` emptyArgTree { _cmd  = "stop"
-                                                        , _ids  = [1]
-                                                        , _opts = ArgOpts True
-                                                        }
+      parseArgs "- 1 --json" `shouldBe` emptyArgTree { _cmd  = "stop"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "toggle expr" $ do
     it "invalid args" $ do
@@ -172,10 +162,10 @@ spec = parallel $ do
     it "with id" $ do
       parseArgs "toggle 1"
         `shouldBe` emptyArgTree { _cmd = "toggle", _ids = [1] }
-      parseArgs "toggle 1 --json" `shouldBe` emptyArgTree { _cmd  = "toggle"
-                                                          , _ids  = [1]
-                                                          , _opts = ArgOpts True
-                                                          }
+      parseArgs "t 1 --json" `shouldBe` emptyArgTree { _cmd  = "toggle"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "done expr" $ do
     it "invalid args" $ do
@@ -184,10 +174,10 @@ spec = parallel $ do
 
     it "with id" $ do
       parseArgs "done 1" `shouldBe` emptyArgTree { _cmd = "done", _ids = [1] }
-      parseArgs "done 1 --json" `shouldBe` emptyArgTree { _cmd  = "done"
-                                                        , _ids  = [1]
-                                                        , _opts = ArgOpts True
-                                                        }
+      parseArgs "d 1 --json" `shouldBe` emptyArgTree { _cmd  = "done"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "delete expr" $ do
     it "invalid args" $ do
@@ -197,10 +187,10 @@ spec = parallel $ do
     it "with id" $ do
       parseArgs "delete 1"
         `shouldBe` emptyArgTree { _cmd = "delete", _ids = [1] }
-      parseArgs "delete 1 --json" `shouldBe` emptyArgTree { _cmd  = "delete"
-                                                          , _ids  = [1]
-                                                          , _opts = ArgOpts True
-                                                          }
+      parseArgs "D 1 --json" `shouldBe` emptyArgTree { _cmd  = "delete"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "remove expr" $ do
     it "invalid args" $ do
@@ -210,10 +200,10 @@ spec = parallel $ do
     it "with id" $ do
       parseArgs "remove 1"
         `shouldBe` emptyArgTree { _cmd = "remove", _ids = [1] }
-      parseArgs "remove 1 --json" `shouldBe` emptyArgTree { _cmd  = "remove"
-                                                          , _ids  = [1]
-                                                          , _opts = ArgOpts True
-                                                          }
+      parseArgs "r 1 --json" `shouldBe` emptyArgTree { _cmd  = "remove"
+                                                     , _ids  = [1]
+                                                     , _opts = ArgOpts True
+                                                     }
 
   describe "context expr" $ do
     it "clear context" $ do
@@ -226,7 +216,7 @@ spec = parallel $ do
         }
 
     it "with opt" $ do
-      parseArgs "ctx +tag1 tag2 --json" `shouldBe` emptyArgTree
+      parseArgs "c +tag1 tag2 --json" `shouldBe` emptyArgTree
         { _cmd  = "context"
         , _tags = ["tag1", "tag2"]
         , _opts = ArgOpts True
@@ -235,24 +225,24 @@ spec = parallel $ do
   describe "list expr" $ do
     it "with opt" $ do
       parseArgs "list  " `shouldBe` emptyArgTree { _type = Qry, _cmd = "list" }
-      parseArgs "list   --json" `shouldBe` emptyArgTree { _type = Qry
-                                                        , _cmd  = "list"
-                                                        , _opts = ArgOpts True
-                                                        }
+      parseArgs "l   --json" `shouldBe` emptyArgTree { _type = Qry
+                                                     , _cmd  = "list"
+                                                     , _opts = ArgOpts True
+                                                     }
 
-  describe "show expr" $ do
+  describe "info expr" $ do
     it "invalid args" $ do
-      parseArgs "show" `shouldBe` emptyArgTree
-      parseArgs "show bad-id" `shouldBe` emptyArgTree
+      parseArgs "info" `shouldBe` emptyArgTree
+      parseArgs "info bad-id" `shouldBe` emptyArgTree
 
     it "with id" $ do
-      parseArgs "show 1"
-        `shouldBe` emptyArgTree { _type = Qry, _cmd = "show", _ids = [1] }
-      parseArgs "show 1   --json" `shouldBe` emptyArgTree { _type = Qry
-                                                          , _cmd  = "show"
-                                                          , _ids  = [1]
-                                                          , _opts = ArgOpts True
-                                                          }
+      parseArgs "info 1"
+        `shouldBe` emptyArgTree { _type = Qry, _cmd = "info", _ids = [1] }
+      parseArgs "i 1   --json" `shouldBe` emptyArgTree { _type = Qry
+                                                       , _cmd  = "info"
+                                                       , _ids  = [1]
+                                                       , _opts = ArgOpts True
+                                                       }
 
   describe "worktime expr" $ do
     it "global" $ do
@@ -275,7 +265,7 @@ spec = parallel $ do
         }
 
     it "with min date" $ do
-      parseArgs "wtime [10 tag" `shouldBe` emptyArgTree
+      parseArgs "w [10 tag" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -283,7 +273,7 @@ spec = parallel $ do
         }
 
     it "with min time" $ do
-      parseArgs "wtime tag [:20" `shouldBe` emptyArgTree
+      parseArgs "w tag [:20" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -291,7 +281,7 @@ spec = parallel $ do
         }
 
     it "with min datetime" $ do
-      parseArgs "wtime [10:20 tag" `shouldBe` emptyArgTree
+      parseArgs "w [10:20 tag" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -299,7 +289,7 @@ spec = parallel $ do
         }
 
     it "with max date" $ do
-      parseArgs "wtime tag ]10" `shouldBe` emptyArgTree
+      parseArgs "w tag ]10" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -307,7 +297,7 @@ spec = parallel $ do
         }
 
     it "with max time" $ do
-      parseArgs "wtime ]:20 tag" `shouldBe` emptyArgTree
+      parseArgs "w ]:20 tag" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -315,7 +305,7 @@ spec = parallel $ do
         }
 
     it "with max datetime" $ do
-      parseArgs "wtime tag ]10:20" `shouldBe` emptyArgTree
+      parseArgs "w tag ]10:20" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]
@@ -323,7 +313,7 @@ spec = parallel $ do
         }
 
     it "with min and max datetimes" $ do
-      parseArgs "wtime [10:20 ]14 tag" `shouldBe` emptyArgTree
+      parseArgs "w [10:20 ]14 tag" `shouldBe` emptyArgTree
         { _type    = Qry
         , _cmd     = "worktime"
         , _tags    = ["tag"]